pyproject.toml
[tool.poetry]
name = "pyhass-zont"
version = "0.1.0"
description = ""
authors = ["kulig <kuligleo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.0"
paho-mqtt = "^2.0.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
ruff = "^0.3.7"
strenum = "^0.4.15"
mypy = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
